Задача: BJS2-67379/ Кэширование свободных хэшей

+++ src/main/java/faang/school/urlshortenerservice/repository/HashRepository.java
@@ -29,5 +29,5 @@ public interface HashRepository extends JpaRepository<Hash, String> {
                      LIMIT :batchSize)
                      RETURNING hash
                      """)
-    List<Hash> getHashBatch(long batch);
+    List<Hash> getHashBatch(int batch);
 }
